+++ src/KitLib.Core/Host/KitLibUserOps.cs
@@ -2,7 +2,6 @@ namespace KitLib.Host;
 
 public static class KitLibUserOps {
     public static Action? OpenLogs { get; set; }
-    public static Action? OpenManual { get; set; }
     public static Func<string?>? CurrentSessionLogFileName { get; set; }
     public static Func<string?>? CurrentSessionLogPath { get; set; }
 }