+++ src/KitLib.Modules.Dev/EnemyIntent/MonsterIntentReader.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json.Nodes;
 using KitLib.AI.Knowledge;
-using KitLib.UI;
 using MegaCrit.Sts2.Core.Combat;
 using MegaCrit.Sts2.Core.Context;
 using MegaCrit.Sts2.Core.Entities.Creatures;